+++ packages/maker/src/api/defineAction.ts
@@ -271,7 +271,10 @@ export function defineCreateObjectAction(
       ?? createDefaultParameterOrdering(def, parameters),
     ...(def.actionLevelValidation
       ? {
-        validation: convertValidationRule(def.actionLevelValidation),
+        validation: convertValidationRule(
+          def.actionLevelValidation,
+          parameters,
+        ),
       }
       : {}),
     ...(def.defaultFormat && { defaultFormat: def.defaultFormat }),
@@ -485,7 +488,10 @@ export function defineModifyObjectAction(
       ),
     ...(def.actionLevelValidation
       ? {
-        validation: convertValidationRule(def.actionLevelValidation),
+        validation: convertValidationRule(
+          def.actionLevelValidation,
+          parameters,
+        ),
       }
       : {}),
     ...(def.defaultFormat && { defaultFormat: def.defaultFormat }),
@@ -541,7 +547,9 @@ export function defineDeleteObjectAction(
     },
     ...(def.actionLevelValidation
       ? {
-        validation: convertValidationRule(def.actionLevelValidation),
+        validation: convertValidationRule(
+          def.actionLevelValidation,
+        ),
       }
       : {}),
   });
@@ -639,7 +647,10 @@ export function defineCreateOrModifyObjectAction(
       ),
     ...(def.actionLevelValidation
       ? {
-        validation: convertValidationRule(def.actionLevelValidation),
+        validation: convertValidationRule(
+          def.actionLevelValidation,
+          parameters,
+        ),
       }
       : {}),
     ...(def.defaultFormat && { defaultFormat: def.defaultFormat }),
@@ -1161,10 +1172,11 @@ function sanitize(s: string): string {
 
 function convertValidationRule(
   actionValidation: ActionLevelValidationDefinition,
+  actionParameters?: ActionParameter[],
 ): Array<ActionValidationRule> {
   return actionValidation.map(rule => {
     return {
-      condition: convertConditionDefinition(rule.condition),
+      condition: convertConditionDefinition(rule.condition, actionParameters),
       displayMetadata: rule.displayMetadata ?? {
         failureMessage: "Did not satisfy validation",
         typeClasses: [],